train and evaluate working
api is WIP
