+++ mentor_classifier/bin/mentor_classifier
@@ -1,0 +1,25 @@
+#!/usr/bin/env python
+import click
+import os
+from mentor_classifier.classifier.train import train
+
+
+@click.group()
+def cli():
+    pass
+
+
+@cli.command()
+@click.option("-p", "--mentor", required=True, type=click.Path(exists=True))
+@click.option("-s", "--shared", required=False, type=click.Path(exists=True))
+@click.option("-o", "--output", required=False, type=click.Path(exists=False))
+def build(mentor: str, shared: str, output: str):
+    scores, accuracy, path = train(mentor_id=mentor, shared_root=shared, output_dir=output)
+    print("Models are saved at: " + path)
+
+
+cli.add_command(build)
+
+
+if __name__ == "__main__":
+    cli()
