--- mentor_classifier/bin/mentor_pipeline
@@ -1,27 +1,0 @@
-#!/usr/bin/env python
-import click
-import os
-from mentor_classifier.classifier import train
-
-
-@click.group()
-def cli():
-    pass
-
-
-@cli.command()
-@click.option("-p", "--mentor", required=False, type=click.Path(exists=True))
-@click.option("-s", "--shared", required=False, type=click.Path(exists=True))
-@click.option("-o", "--output", required=False, type=click.Path(exists=False))
-def train(mentor: str, shared: str, output: str):
-    result = train(mentor_id=mentor, shared_root=shared, output_dir=output)
-    print("Models are saved at: " + output)
-    for exp_num, exp_result in enumerate(result.questions):
-        print(f"Accuracy for model={exp_num} is {exp_result.accuracy}.")
-
-
-cli.add_command(train)
-
-
-if __name__ == "__main__":
-    cli()
