+++ mentor_classifier_api/Makefile
@@ -1,8 +1,8 @@
 BLACK_EXCLUDES="/(\.venv|build)/"
-DOCKER_IMAGE?=mentor-pipeline-api
+DOCKER_IMAGE?=mentor-classifier-api
 DOCKER_IMAGE_ID=$(shell docker images -q ${DOCKER_IMAGE} 2> /dev/null)
 ROOT=$(shell dirname ${PWD})
-MENTOR_PIPELINE_SRC=$(ROOT)/mentor_classifier
+MENTOR_CLASSIFIER_SRC=$(ROOT)/mentor_classifier
 VENV=.venv
 VENV_PIP=$(VENV)/bin/pip
 $(VENV):
@@ -18,10 +18,7 @@ $(VENV)-update: $(VENV)-installed
 	$(VENV_PIP) install --upgrade pip
 	$(VENV_PIP) install -r requirements.test.txt
 	$(VENV_PIP) install ../mentor_classifier
-	$(VENV)/bin/python3.6 -m nltk.downloader punkt
-	$(VENV)/bin/python3.6 -m nltk.downloader wordnet
 	$(VENV)/bin/python3.6 -m nltk.downloader averaged_perceptron_tagger
-	$(VENV)/bin/python3.6 -m nltk.downloader stopwords
 
 .PHONY clean:
 clean:
@@ -35,7 +32,7 @@ build/mentor_classifier:
 			--exclude .pytest_cache \
 			--exclude tests \
 			--exclude __pycache__ \
-		$(MENTOR_PIPELINE_SRC) build
+		$(MENTOR_CLASSIFIER_SRC) build
 	
 .PHONY docker-build:
 docker-build: clean build/mentor_classifier
