+++ mentor_classifier_api/tests/fixtures/models/clint/clint/w2v.txt
@@ -1,0 +1,1 @@
+/Users/kcarr/Documents/mentor-admin-classifier/word2vec/word2vec.bin
