--- mentor_classifier_api/tests/fixtures/models/q1/build_version.json
@@ -1,7 +1,0 @@
-{
-    "pipeline": "1.0.0",
-    "pipelineApi": "1.0.0",
-    "modelPIPELINE": "1.0.0",
-    "modelId": "q1",
-    "modelUpdatedAt": "2020-08-04T11:45:00Z"
-}
