--- mentor_classifier_api/tests/fixtures/models/q1/config.yaml
@@ -1,26 +1,0 @@
-##
-## This software is Copyright ©️ 2020 The University of Southern California. All Rights Reserved. 
-## Permission to use, copy, modify, and distribute this software and its documentation for educational, research and non-profit purposes, without fee, and without a written agreement is hereby granted, provided that the above copyright notice and subject to the full license file found in the root of this software deliverable. Permission to make commercial use of this software may be obtained by contacting:  USC Stevens Center for Innovation University of Southern California 1150 S. Olive Street, Suite 2300, Los Angeles, CA 90115, USA Email: accounting@stevens.usc.edu
-##
-## The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
-##
-question: What are the challenges to demonstrating integrity in a group?
-questions:
-- ideal: peer pressure cause allow inappropriate behavior
-  features:
-    bad:
-      - \brule\w*\b.*\bdiffer\w*\b
-    good:
-      - \b(group|peer|lead\w*|command\w*)\b.*\b(pressu\w*|influ\w*)\b
-- ideal: may get trouble
-  features:
-    bad:
-      - \b(lov\w*|lik\w*|car)\b
-    good:
-      - \b(troub\w*|hard(er)|embarrass(ed|ing)|defens\w*|ang\w*|ups\w*)\b
-- ideal: enforcing rule make unpopular
-  features:
-    bad:
-      - \b(super\w*|bette\w*)\b
-    good:
-      - \b(unpop\w*|hate\w*|unlik\w*|not\040popular)\b
